ilasm: Some work on data constants.
Note that this implementation is far from standard-compliant. At the
moment, we just copy the data constant's value into the field's
InitialValue and nothing else. The tls/cil modifiers are ignored,
too. Even if we did implement data constants properly, the Mono
runtime doesn't actually handle them the same way as the Microsoft
runtime, so we'd be incompatible anyway.