ilasm: Convert to deferred type resolution.
With this approach, types are evaluated 'lazily', i.e. when we need
them instead of right when we encounter them. This approach helps in
resolution of generic parameters because we can delay full resolution
of GPs till we have the full context to resolve them in.

Idea proposed by Jb - thanks!

Also fixed a couple of random bugs along the way.

(Important side-note: Do not reference $n inside a closure, as this
 will result in accessing the state array later, leading to incorrect
 values being returned.)